[Development Branch] [Code Changes] [Santiago Arellano]
The present commit includes a series of modifications that allow for CSV formatting, parsing and serializing based on multilevel composition between the main backend classes (Employee and Manager) classes, and a series of new custom Exceptions. Through these classes, the goal of serializing by CSV has been achieved and is in need of testing.