[Development Branch][Bug Fixes And Changes] [Santiago Arellano]
The present commit, brings changes into most of the backend for the application. All the serialization methods have been written, tested, and are awaiting connection with the UI once the salary calculation methods have been polished and the field constants that will be used are done. Changes into the way most classes hold the data for the Secondary Title that an employee can hold has been altered in the Manager class. The introduction of three new constants has allowed for a streamlined environment which will guarantee a correct input, output, display and modification of data through the UI and the serialized versions of the EmployeeListWrapper.

For the JSON serialization method, while constant data members with the names of the fields have not been added, they remain constant through both manager and employee instances. The only difference is the Key that points to the entire collection of data (This key could be Manager or Employee but not both). For XML, the structure has been defined polymorphic throughout the codebase for the EmployeeAbstraction package. This allows for serialization of both classes inside the same file, and deserialization from the same file.

In terms of CSV serialization, a more robust check for the Titulo of Manager has been added, serialization is more standard using the aforementioned constants as well as the toCSVString() methods defined in the EmployeeAbstraction package classes. Finally, as for Binary, both serialization and deserialization have been added and work flawlessly given their integration with the Serializable Annotation interface.

Overall functionality can be guaranteed. However, cohesion with the Ui cannot be guaranteed at this time as the methods, most important ones, have not been developed yet, as such, most of the integration with the TableViews, sorting, editing, and filtering have not been developed as of yet.
Tested-by: Santiago Arellano <santiagofarellanoj@icloud.com>