Final Commit.
The Project has been finished, despite the setbacks, the difficulties and how utterly massive the codebase has become, the project is finished and ready for presentation. All the methods work, all functionality for exporting and importing works, visualization of data, addition of data, modification, deletion, and most importantly of all visualizations are fully integrated and fully fledged. All serialization methods, exporting and calculation methods have been implemented. The application is robust, and I have yet to notice any error message flash in the console or unexpected behavior to happen.

Tested-by: Santiago Arellano <santiagofarellanoj@icloud.com>